Module bl602_hal::prelude

HAL crate prelude

Re-exports

pub use crate::gpio::GlbExt as _bl602_hal_gpio_GlbExt;

Traits

Extensions
Create rate-based extensions from primitive numeric types.

_embedded_hal_Capture
Input capture

_embedded_hal_Pwm
Pulse Width Modulation

_embedded_hal_PwmPin
A single PWM channel / pin

_embedded_hal_Qei
Quadrature encoder interface

_embedded_hal_adc_Channel
A marker trait to identify MCU pins that can be used as inputs to an ADC channel.

_embedded_hal_adc_OneShot
ADCs that sample on single channels per request, and do so at the time of the request.

_embedded_hal_blocking_delay_DelayMs
Millisecond delay

_embedded_hal_blocking_delay_DelayUs
Microsecond delay

_embedded_hal_blocking_i2c_Read
Blocking read

_embedded_hal_blocking_i2c_Transactional
Transactional I2C interface.

_embedded_hal_blocking_i2c_Write
Blocking write

_embedded_hal_blocking_i2c_WriteIter
Blocking write (iterator version)

_embedded_hal_blocking_i2c_WriteIterRead
Blocking write (iterator version) + read

_embedded_hal_blocking_i2c_WriteRead
Blocking write + read

_embedded_hal_blocking_rng_Read
Blocking read

_embedded_hal_blocking_serial_Write
Write half of a serial interface (blocking variant)

_embedded_hal_blocking_spi_Transfer
Blocking transfer

_embedded_hal_blocking_spi_Write
Blocking write

_embedded_hal_blocking_spi_WriteIter
Blocking write (iterator version)

_embedded_hal_digital_InputPin
Single digital input pin

_embedded_hal_digital_OutputPin
Single digital push-pull output pin

_embedded_hal_digital_StatefulOutputPin
Push-pull output pin that can read its output state

_embedded_hal_digital_ToggleableOutputPin
Output pin that can be toggled

_embedded_hal_rng_Read
Nonblocking stream of random bytes.

_embedded_hal_serial_Read
Read half of a serial interface

_embedded_hal_serial_Write
Write half of a serial interface

_embedded_hal_spi_FullDuplex
Full duplex (master mode)

_embedded_hal_timer_Cancel
Trait for cancelable countdowns.

_embedded_hal_timer_CountDown
A count down timer

_embedded_hal_timer_Periodic
Marker trait that indicates that a timer is periodic

_embedded_hal_watchdog_Disable
Disables a running watchdog timer so the processor won't be reset.

_embedded_hal_watchdog_Enable
Enables A watchdog timer to reset the processor if software is frozen or stalled.

_embedded_hal_watchdog_Watchdog
Feeds an existing watchdog to ensure the processor isn't reset. Sometimes the "feeding" operation is also referred to as "refreshing".