Trait bl602_hal::prelude::_embedded_hal_digital_OutputPin

pub trait _embedded_hal_digital_OutputPin {
    type Error;
    fn try_set_low(&mut self) -> Result<(), Self::Error>;

    fn try_set_high(&mut self) -> Result<(), Self::Error>;

    fn try_set_state(&mut self, state: PinState) -> Result<(), Self::Error> { ... }
}

Single digital push-pull output pin

Associated Types

type Error

Error type

Loading content...
Required methods

fn try_set_low(&mut self) -> Result<(), Self::Error>

Drives the pin low

NOTE the actual electrical state of the pin may not actually be low, e.g. due to external electrical sources

fn try_set_high(&mut self) -> Result<(), Self::Error>

Drives the pin high

NOTE the actual electrical state of the pin may not actually be high, e.g. due to external electrical sources

Loading content...
Provided methods

fn try_set_state(&mut self, state: PinState) -> Result<(), Self::Error>

Drives the pin high or low depending on the provided value

NOTE the actual electrical state of the pin may not actually be high or low, e.g. due to external electrical sources

Loading content...