Trait bl602_hal::prelude::_embedded_hal_digital_StatefulOutputPin

pub trait _embedded_hal_digital_StatefulOutputPin: OutputPin {
    fn try_is_set_high(&self) -> Result<bool, Self::Error>;

    fn try_is_set_low(&self) -> Result<bool, Self::Error>;
}

Push-pull output pin that can read its output state

Required methods

fn try_is_set_high(&self) -> Result<bool, Self::Error>

Is the pin in drive high mode?

NOTE this does not read the electrical state of the pin

fn try_is_set_low(&self) -> Result<bool, Self::Error>

Is the pin in drive low mode?

NOTE this does not read the electrical state of the pin

Loading content...